+++ src/render/rendermodes/base.h
@@ -49,9 +49,9 @@ public:
 	virtual void end();
 
 	// is called to allow the rendermode to hide specific blocks
-	virtual bool isHidden(const mc::BlockPos& pos, uint16_t id, uint8_t data);
+	virtual bool isHidden(const mc::BlockPos& pos, uint16_t id, uint16_t data);
 	// is called to allow the rendermode to change a block image
-	virtual void draw(Image& image, const mc::BlockPos& pos, uint16_t id, uint8_t data);
+	virtual void draw(Image& image, const mc::BlockPos& pos, uint16_t id, uint16_t data);
 };
 
 bool createRendermode(const std::string& name, const RenderWorldConfig& config,
